+++ packages/sources/finage-test/src/endpoint/ws/forex-ws.ts
@@ -1,6 +1,6 @@
 import { WebSocketTransport } from '@chainlink/external-adapter-framework/transports/websocket'
 import { SingleNumberResultResponse } from '@chainlink/external-adapter-framework/util'
-import { customSettings } from '../../config'
+import { config } from '../../config'
 import { ForexEndpointParams } from '../forex-router'
 
 interface Message {
@@ -14,7 +14,7 @@ type EndpointTypes = {
     Params: ForexEndpointParams
   }
   Response: SingleNumberResultResponse
-  CustomSettings: typeof customSettings
+  Settings: typeof config.settings
   Provider: {
     WsMessage: Message
   }
@@ -22,7 +22,7 @@ type EndpointTypes = {
 
 export const wsTransport = new WebSocketTransport<EndpointTypes>({
   url: (context) => {
-    return `${context.adapterConfig.FOREX_WS_API_ENDPOINT}/?token=${context.adapterConfig.WS_SOCKET_KEY}`
+    return `${context.adapterSettings.FOREX_WS_API_ENDPOINT}/?token=${context.adapterSettings.WS_SOCKET_KEY}`
   },
   handlers: {
     message(message) {