+++ packages/sources/tiingo-test/src/endpoint/http/vwap.ts
@@ -1,9 +1,9 @@
-import { HttpTransport } from '@chainlink/external-adapter-framework/transports'
-import { constructEntry, CryptoEndpointTypes, inputParameters } from '../../crypto-utils'
 import { PriceEndpoint } from '@chainlink/external-adapter-framework/adapter'
+import { HttpTransport } from '@chainlink/external-adapter-framework/transports'
 import overrides from '../../config/overrides.json'
+import { constructEntry, HttpTransportTypes, inputParameters } from '../../crypto-utils'
 
-export const httpTransport = new HttpTransport<CryptoEndpointTypes>({
+export const httpTransport = new HttpTransport<HttpTransportTypes>({
   prepareRequests: (params, config) => {
     return params.map((param) => {
       return {
@@ -27,7 +27,7 @@ export const httpTransport = new HttpTransport<CryptoEndpointTypes>({
   },
 })
 
-export const endpoint = new PriceEndpoint<CryptoEndpointTypes>({
+export const endpoint = new PriceEndpoint({
   name: 'vwap',
   aliases: ['crypto-vwap'],
   transport: httpTransport,