+++ packages/composites/crypto-volatility-index/CHANGELOG.md
@@ -1,5 +1,23 @@
 # @chainlink/crypto-volatility-index-adapter
 
+## 1.3.40
+
+### Patch Changes
+
+- @chainlink/token-allocation-adapter@1.9.6
+
+## 1.3.39
+
+### Patch Changes
+
+- @chainlink/token-allocation-adapter@1.9.5
+
+## 1.3.38
+
+### Patch Changes
+
+- @chainlink/token-allocation-adapter@1.9.4
+
 ## 1.3.37
 
 ### Patch Changes