+++ packages/composites/curve-3pool/CHANGELOG.md
@@ -1,5 +1,23 @@
 # @chainlink/curve-3pool-adapter
 
+## 3.0.21
+
+### Patch Changes
+
+- @chainlink/token-allocation-adapter@1.9.6
+
+## 3.0.20
+
+### Patch Changes
+
+- @chainlink/token-allocation-adapter@1.9.5
+
+## 3.0.19
+
+### Patch Changes
+
+- @chainlink/token-allocation-adapter@1.9.4
+
 ## 3.0.18
 
 ### Patch Changes