+++ packages/composites/defi-dozen/CHANGELOG.md
@@ -1,5 +1,23 @@
 # @chainlink/defi-dozen-adapter
 
+## 1.2.40
+
+### Patch Changes
+
+- @chainlink/token-allocation-adapter@1.9.6
+
+## 1.2.39
+
+### Patch Changes
+
+- @chainlink/token-allocation-adapter@1.9.5
+
+## 1.2.38
+
+### Patch Changes
+
+- @chainlink/token-allocation-adapter@1.9.4
+
 ## 1.2.37
 
 ### Patch Changes