+++ packages/composites/linear-finance/CHANGELOG.md
@@ -1,5 +1,23 @@
 # @chainlink/linear-finance-adapter
 
+## 2.3.29
+
+### Patch Changes
+
+- @chainlink/token-allocation-adapter@1.9.6
+
+## 2.3.28
+
+### Patch Changes
+
+- @chainlink/token-allocation-adapter@1.9.5
+
+## 2.3.27
+
+### Patch Changes
+
+- @chainlink/token-allocation-adapter@1.9.4
+
 ## 2.3.26
 
 ### Patch Changes