+++ packages/composites/outlier-detection/CHANGELOG.md
@@ -1,5 +1,23 @@
 # @chainlink/outlier-detection-adapter
 
+## 1.2.72
+
+### Patch Changes
+
+- @chainlink/ea@1.3.71
+
+## 1.2.71
+
+### Patch Changes
+
+- @chainlink/ea@1.3.70
+
+## 1.2.70
+
+### Patch Changes
+
+- @chainlink/ea@1.3.69
+
 ## 1.2.69
 
 ### Patch Changes