+++ packages/composites/savax-price/CHANGELOG.md
@@ -1,5 +1,23 @@
 # @chainlink/savax-price-adapter
 
+## 3.0.22
+
+### Patch Changes
+
+- @chainlink/token-allocation-adapter@1.9.6
+
+## 3.0.21
+
+### Patch Changes
+
+- @chainlink/token-allocation-adapter@1.9.5
+
+## 3.0.20
+
+### Patch Changes
+
+- @chainlink/token-allocation-adapter@1.9.4
+
 ## 3.0.19
 
 ### Patch Changes